RFC: avx512/ternarylogic: two versions with many logic reductions
I have deliberately left a #if 0/1 to switch between my two best versions (so far ;-) ).
They both use exactly the same algorithm except that #if 1 uses SIMDE_VECTORIZE (it suffers from the clang imm8 = 88 issue)
and #if 0 uses explicit SIMDe SSE2 intrinsics which works around this problem (but the emul version is very slow to compile).
I am pretty sure that it will be possible to introduce a special case in the SIMDE_VECTORIZE version
to work around the clang 88 problem and hopefully any other similar problems.